[Client ]  Update Login, Signup feature

+++ server/controllers/users/delete.js
@@ -1,14 +1,18 @@
 const { user } = require("../../models");
+const { isAuthorized } = require("../tokenFunctions");
 
-module.exports = (res, req) => {
-  const userInfo = user.destroy({
-    where: { id: 1 },
-  });
+module.exports = async (res, req) => {
+  const userinfo = isAuthorized(req);
+  // TODO: 로그인 여부를 판단하고, Access token payload를 이용하여 응답을 제공하세요.
+  console.log("accessTokenData : ", accessTokenData);
+  // isAuthorized함수에서 데이터를 잘 받아왔으면 accessTokenData에 유저정보가 담겨있고 그렇지 않으면 마찬가지로 undefined가 담겨있음.
 
-  console.log(userInfo);
-
-  res
-    // .cookie("jwt", null, { maxAge: 0 })
-    .status(200)
-    .json({ id: 1 });
+  if (!userinfo) {
+    res.status(401).send({ message: "Not Authorized" });
+  } else {
+    await user.destroy({
+      where: { id: userinfo.id },
+    });
+    res.status(200).send({ message: "User Delete Ok" });
+  }
 };
